Give the syntax item flags a documented home
The `HL_*` syntax flags (`HL_CONTAINED`, `HL_TRANSP`, `HL_ONELINE`, ...)
were the last of the flag families this batch was asked to consolidate,
in the shape B12-4 used for `HLF_*`. Unlike `HLF_*` they were not
actually duplicated -- all twenty already lived in syntax.rs -- so the
job here is the other half of that treatment: a real home, a real type,
and a doc comment each.

They arrived as a c2rust anonymous-enum block: twenty
`pub const HL_X: C2Rust_Unnamed_23 = <decimal>` with the alias typed
`c_uint`, sitting between two struct definitions. Every one of the three
carriers (`synpat_T::sp_flags`, `keyentry::flags`, `stateitem_T::
si_flags`) is a `c_int`, so each of the 116 uses tree-wide spelled out
`as ::core::ffi::c_int`. Typing the constants to the consumer deletes
all of them.

`syntax/flags.rs` is `forbid(unsafe_code)` and carries the enum's own
`v0.12.4:src/nvim/syntax.h` documentation, plus the two facts that
reading the values does not tell you: which of the three carriers each
flag can appear on (`HL_MATCHCONT`/`HL_TRANS_CONT` are state-machine
deductions with no `:syntax` keyword behind them), and that `HL_CONCEAL`
here is *not* `highlight::HL_CONCEALED` -- the two families share a
prefix and mean unrelated things.

Values re-checked against the C enum one by one, and written back as the
hex the header uses.

+++ crates/nvim/src/nvim/eval/funcs/screen.rs
@@ -416,7 +416,7 @@ pub unsafe extern "C" fn f_synconcealed(
             // reports on the position it last looked at.
             syn_get_id(curwin.get(), lnum, col, false_0, ptr::null_mut(), false_0);
             syntax_flags = get_syntax_info(&raw mut matchid);
-            if syntax_flags & HL_CONCEAL as c_int != 0 && (*curwin.get()).w_onebuf_opt.wo_cole < 3 {
+            if syntax_flags & HL_CONCEAL != 0 && (*curwin.get()).w_onebuf_opt.wo_cole < 3 {
                 let mut cchar = schar_from_char(syn_get_sub_char());
                 // At 'conceallevel' 1 a group with no `cchar` falls back to
                 // 'listchars' "conceal", and to a space if that is unset.
@@ -435,7 +435,7 @@ pub unsafe extern "C" fn f_synconcealed(
         let list = tv_list_alloc_ret(rettv, 3);
         tv_list_append_number(
             list,
-            (syntax_flags & HL_CONCEAL as c_int != 0) as c_int as varnumber_T,
+            (syntax_flags & HL_CONCEAL != 0) as c_int as varnumber_T,
         );
         tv_list_append_string(list, text.as_ptr(), -1);
         tv_list_append_number(list, matchid as varnumber_T);
